fix ast code + more
fixed ast code to use recursive logic instead of loop. The ast function also returns the addr of the root node now instead of the illogical "end" addr. This will enabling folding over the ast in the next step by jumping through the ast tree which can't and isn't laid out linearly in memory.

There is also a new formatting thing where (;;) will appear before an expression to indicate that it leaves a value on the stack to either be returned at function end or consumed by a future expression.